Extract command hashmap to parser class
